refactor: split phase3/profile.rs further (types/run/sweep/summary)
profile.rs at 967 lines was still F-grade (0.0 TDG Structural).
Split into sub-submodules under phase3/profile/: types, sweep, run,
summary. All files now A- (92.6–100 TDG Structural).

Refs paiml/infra#PMAT-157
